requirements.txt
crewai
crewai-tools
streamlit
fastapi
uvicorn
pydantic
python-dotenv
youtube-transcript-api 